% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_granges.R
\name{convert_granges}
\alias{convert_granges}
\title{GRanges object conversion}
\usage{
convert_granges(obj, ...)
}
\arguments{
\item{obj}{number of regions to create}

\item{...}{arguments passed to \code{makeGRangesFromDataFrame}}
}
\value{
if obj is character vector or data.frame, it is converted to a GRanges object. If it is a GRanges object, it is converted to a character vector
}
\description{
convert GRanges to strings or data.frame/strings to GRanges
}
\examples{
region_vec = c("chr1-105442114-105442778", "chr1-105453942-105454629", "chr1-106231358-106231675")
#convert region strings to GRanges
(region_gr = convert_granges(region_vec))
#convert GRanges to region strings
convert_granges(region_gr)
#convert data.frame to GRanges
region_df = data.frame(seqnames=c('chr1', 'chr2', 'chr2'),
                       start=c(1,100,345),
                       end = c(306, 3409,909),
                       strand = c('*','+', '-'), 
                       symbol=c('gene_1', 'gene_2', 'gene_3'))
convert_granges(region_df, keep.extra.columns = TRUE)
}
