% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentage_overlap.R
\name{get_percentage_overlap}
\alias{get_percentage_overlap}
\alias{get_signature_overlap_archr}
\alias{get_signature_overlap_seurat}
\title{percentage signature overlap}
\usage{
get_percentage_overlap(
  peak_matrix,
  reduced_dim_df,
  signature_gr,
  nFrags_vec,
  count_thres = 3e+05,
  k = 100,
  min_overlap = 0.5,
  workers = 1
)

get_signature_overlap_archr(
  proj,
  signature_gr,
  reduction = "UMAP",
  count_thres = 3e+05,
  k = 100,
  min_overlap = 0.5
)

get_signature_overlap_seurat(
  seurat_obj,
  signature_gr,
  nFrags_vec,
  assay = "scATAC_raw",
  slot = "counts",
  reduction = "umap",
  count_thres = 3e+05,
  k = 100,
  min_overlap = 0.5
)
}
\arguments{
\item{peak_matrix}{binary matrix with peak regions in rows (format chr1-1-99) and cells in columns}

\item{reduced_dim_df}{data.frame with cells in rows and coordinates from dimension reduction in columns}

\item{signature_gr}{peak signature to quantify as GRanges object}

\item{nFrags_vec}{vector with the fragment count per cell (same length as ncol in peak matrix)}

\item{count_thres}{number of fragments to aggregate cells to (nearest neighbour counts are added until the threshold is reached)}

\item{k}{number of nearest neighbours to search for each cell}

\item{min_overlap}{minimum overlap of the peak regions with signature regions to count as overlapping feature}

\item{workers}{number of workers to use}

\item{proj}{ArchR project to use}

\item{reduction}{dimensionality reduction from ArchR proj, which is used to search nearest neighbours}
}
\description{
For each cell calculate how high the percentage of signature peaks covered by the peaks of that cell is.
To avoid bias due to count depth, perform an aggregation to above a count threshold among nearest neighbours
}
