% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature_pvalue.R
\name{get_signature_pvalue}
\alias{get_signature_pvalue}
\title{signature enrichment/depletetion p-value}
\usage{
get_signature_pvalue(
  pmat,
  group_vector,
  signature_peaks_gr,
  genome,
  n_background_sets = 300
)
}
\arguments{
\item{pmat}{binary peak matrix (rows = regions in the format 'chr1-10-100', columns = cells)}

\item{group_vector}{vector of group annotation for the cells in the peak matrix (e.g. clustering result)}

\item{signature_peaks_gr}{GRanges object of the peak signature}

\item{genome}{genome for selection of background sequences, e.g. BSgenome.Mmusculus.UCSC.mm10}

\item{n_background_sets}{number of background sets to compare against, higher values give more precise estimates}
}
\description{
Given a peak signature, (binary) scATAC peakmatrix and a cell-to-cluster annotation, calculate an enrichment or
depletion p-value for each cluster using bias-matched background peak sets
}
\examples{
#bad example since there are no real clusters in this dataset, but the function still works
signac_atac = Signac::atac_small
peakmat = BinarizeCounts(GetAssayData(signac_atac))
library(BSgenome.Hsapiens.UCSC.hg38)
sig_gr = Signac::StringToGRanges(rownames(FindMarkers(signac_atac, assay = 'peaks', ident.1 = '0')))
res = get_signature_pvalue(pmat = peakmat,
                     group_vector = signac_atac$seurat_clusters, 
                     signature_peaks_gr = sig_gr,
                     genome = BSgenome.Hsapiens.UCSC.hg38)
}
\keyword{ggplot,plot_grobs,plot_grid}
