% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_peak_matrices.R
\name{unify_peak_matrices}
\alias{unify_peak_matrices}
\title{reduce two peak matrices to a common feature set}
\usage{
unify_peak_matrices(pmat_list, minOverlap = 0)
}
\arguments{
\item{pmat_list}{list containing two peak matrices, where each matrix has region strings (e.g. chr-110-324) as rows and samples as columns}

\item{minOlap}{minimum overlap of regions in both directions required to keep them as a common feature}
}
\value{
list with the two peak matrices, but with identical features as rows
}
\description{
Filters two peak matrices for their overlapping features and returns them with new feature names,
which are identical between the two matrices.
}
\examples{
#get somee random regions as features
set.seed(1)
gr1 = random_granges(n = 30, chr_annot = structure(5000, names='chr1'), width = sample(1:100,30,replace=T))
gr2 = random_granges(n = 20, chr_annot = structure(5000, names='chr1'), width = sample(1:100,30,replace=T))
# show the overlap between the two granges
granges_overlap(gr1, gr2, return_type = 'message')
# intialize peak matrices
pmat1 = matrix(1:90, nrow=30, dimnames=list(convert_granges(gr1), paste0('dataset1_cell_', 1:3)))
pmat2 = matrix(1:40, nrow=20, dimnames=list(convert_granges(gr2), paste0('dataset2_cell_', 1:2)))
# return the matrices with unified features
unify_peak_matrices(list(pmat1, pmat2))
}
