% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_granges.R
\name{random_granges}
\alias{random_granges}
\title{generate random GRanges}
\usage{
random_granges(n, txdb, width = 1)
}
\arguments{
\item{n}{number of regions to create}

\item{txdb}{txdb object containing the sizes of chromosomes}

\item{width}{sizes of the regions, vector of integers from which is sampled}
}
\description{
create random granges objects based on supplied chromosome lenghts and desired region widths
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb = TxDb.Hsapiens.UCSC.hg19.knownGene
random_granges(n = 30, txdb = txdb, width = pmax(1, as.integer(rnorm(30, mean = 50, sd = 30))))
}
